Made the first version of a tutorial. Activates on every start-up.
